+++ src/index.ts
@@ -10,12 +10,23 @@ async function main(argv: string[]): Promise<void>
 {
 	const options = parseArguments(argv);
 	const stream = findEntries(options);
+	let tasks: Promise<number>[] = [];
 	for await(const entry of stream)
 	{
 		for(const exec of options.exec)
 		{
-			await execCommand(exec, entry.toString(), options.startingPoint);
+			tasks.push(execCommand(exec, entry.toString(), options.startingPoint));
+			if(tasks.length >= options.concurrency)
+			{
+				await Promise.all(tasks);
+				tasks = [];
+			}
 		}
+	}
+
+	if(tasks.length > 0)
+	{
+		await Promise.all(tasks);
 	}
 }